pythonanywhere_requirements.txt
flask==2.3.3
requests==2.31.0
telethon==1.40.0
python-dotenv==1.0.0
gunicorn==21.2.0